refactor: refactore to validate reference table
